web/requirements.txt
flask==2.0.3
werkzeug==2.0.3
psycopg2-binary==2.9.3
pandas==1.3.5
numpy==1.21.6